+++ package.json
@@ -40,7 +40,9 @@
   },
   "dependencies": {
     "js-yaml": "^3.6.1",
-    "lodash.isequal": "^4.5.0"
+    "lodash.isequal": "^4.5.0",
+    "remark": "^8.0.0",
+    "traverse": "^0.6.6"
   },
   "jest": {
     "testRegex": "/src/__tests__/.*\\.test.js$"
