+++ src/parser/section/content.js
@@ -1,0 +1,5 @@
+import remark from 'remark'
+
+export default function parseContent (lines, startLineNum, endLineNum) {
+  return remark.parse(lines.slice(startLineNum, endLineNum + 1)).children
+}
